Add StrategyResult type and utility functions for workflow execution
- Introduced a new `StrategyResult` type to standardize the output of workflow execution strategies.
- Added utility function `toExecutionModel` to convert `EsWorkflow` to `WorkflowExecutionEngineModel`, enhancing code reuse.
- Updated `WorkflowExecuteAsyncStrategy` and `WorkflowExecuteSyncStrategy` to utilize the new type and utility function for improved clarity and error handling.
- Created new `types.ts` and `utils.ts` files to organize type definitions and utility functions related to workflow execution.
