+++ src/platform/plugins/shared/workflows_execution_engine/server/step/workflow_execute_step/utils.ts
@@ -7,7 +7,18 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export { WorkflowExecuteStepImpl } from './workflow_execute_step_impl';
-export { WorkflowExecuteAsyncStrategy } from './strategies/workflow_execute_async_strategy';
-export { WorkflowExecuteSyncStrategy } from './strategies/workflow_execute_sync_strategy';
-export type { StrategyResult } from './types';
+import type { EsWorkflow, WorkflowExecutionEngineModel } from '@kbn/workflows';
+
+export function toExecutionModel(
+  workflow: EsWorkflow,
+  isTestRun: boolean
+): WorkflowExecutionEngineModel {
+  return {
+    id: workflow.id,
+    name: workflow.name,
+    enabled: workflow.enabled,
+    definition: workflow.definition,
+    yaml: workflow.yaml,
+    isTestRun,
+  };
+}
